Convert subcomponent screen to Compose
